Added CryptoTracker and BitcoinHistoricalDataScraper
Added a simple CryptoTracker that tracks the ticker of a specific Crypto, sells and buys, and text messages a person when it does sell.

Added a webscraper that combs all information about the historical price data of Bitcoin and puts it into a CSV file to review/use Data Analysis on.